pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mira"
version = "0.1.0"
description = "Production-style learning assistant backend with OpenAI-compatible API"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Mira Contributors" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "scikit-learn>=1.5.0",
  "pandas>=2.2.0",
  "numpy>=2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]
llm = [
  "pyyaml>=6.0.1",
  "huggingface-hub>=0.25.0",
  "datasets>=3.0.0",
  "transformers>=4.45.0",
  "peft>=0.13.0",
  "accelerate>=1.0.0",
  "safetensors>=0.4.5",
  "sentencepiece>=0.2.0"
]
gpu = [
  "bitsandbytes>=0.43.0; platform_system == 'Linux'",
  "flash-attn>=2.6.3; platform_system == 'Linux'"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
